requirements.txt
python-socketio==5.6.0
fastapi==0.75.2
uvicorn==0.17.6
websockets==10.3
attrs==21.4.0
